+++ CHANGELOG.md
@@ -5,16 +5,15 @@ This change log covers only the client part of Sprotty. See also the change logs
 ### v0.10.0 (upcoming)
 
 New features:
- * Line jumps to visually clarify intersecting lines ([#226](https://github.com/eclipse/sprotty/pull/226))
+ * Line jumps (`JumpingPolylineEdgeView`) or gaps (`PolylineEdgeViewWithGapsOnIntersections`) to visually clarify intersecting edges ([#226](https://github.com/eclipse/sprotty/pull/226))
 
  * `TYPES.IEdgeRoutePostprocessor` can be registered to analyse and/or change computed routes ([#226](https://github.com/eclipse/sprotty/pull/226))
 
- * `EdgeRouterRegistry` can route all edges contained in a parent element at once. These pre-computed routes are then usually added to the `args` that are passed on to the views. This allows `IView` and `IEdgeRoutePostprocessor` implementations to consider all computed routes before the routed edges have been rendered. ([#226](https://github.com/eclipse/sprotty/pull/226))
+ * `EdgeRouterRegistry` can route all edges contained in a parent element at once. These pre-computed routes are then added to the `args` that are passed on to the views. This allows `IView` and `IEdgeRoutePostprocessor` implementations to consider all computed routes before the routed edges have been rendered. ([#226](https://github.com/eclipse/sprotty/pull/226))
 
 Breaking API changes:
  * It is recommended that implementations of the `IView` for `SGraph` instances compute the routes of its children with `edgeRouterRegistry.routeAllChildren(model)` and pass on the routes as arguments to its child views. See implementation of `SGraphView` ([#226](https://github.com/eclipse/sprotty/pull/226))
 
- * It is recommended that all implementations of `IView` pass the received `args` on to its child views via `renderChildren()` so that, e.g., pre-computed routes are available to them by default ([#226](https://github.com/eclipse/sprotty/pull/226))
 Breaking API changes:
 
 * Upgrade to snabbdom 3.0.3. The imports of snabbdom functions have changed. The main snabbdom package exports all of the public API.This means consumers of the snabbdom package need to update their imports.